+++ frontends/rioterm/src/context.rs
@@ -392,15 +392,20 @@ impl<T: EventListener + Clone + std::marker::Send + 'static> ContextManager<T> {
     }
 
     #[inline]
-    pub fn schedule_render(&mut self, scheduled_time: u64) {
+    pub fn schedule_update(&mut self) {
         self.event_proxy
-            .send_event(RioEvent::PrepareRender(scheduled_time), self.window_id);
+            .send_event(RioEvent::ProcessUpdate, self.window_id);
+    }
+
+    #[inline]
+    #[allow(unused)]
+    pub fn schedule_render(&mut self) {
+        self.event_proxy
+            .send_event(RioEvent::RenderRoute(self.current_route), self.window_id);
     }
 
     #[inline]
     pub fn blink_cursor(&mut self, scheduled_time: u64) {
-        // PrepareRender will force a render for any route that is focused on window
-        // PrepareRenderOnRoute only call render function for specific route ids.
         self.event_proxy.send_event(
             RioEvent::BlinkCursor(scheduled_time, self.current_route),
             self.window_id,