+++ frontends/rioterm/src/renderer/mod.rs
@@ -16,7 +16,7 @@ use rio_backend::config::colors::{
 };
 use rio_backend::config::Config;
 use rio_backend::sugarloaf::{
-    Content, FragmentStyle, FragmentStyleDecoration, Graphic, Stretch, Style,
+    Content, ContentState, Object, FragmentStyle, FragmentStyleDecoration, Graphic, Stretch, Style,
     SugarCursor, Sugarloaf, UnderlineInfo, UnderlineShape, Weight,
 };
 use std::collections::HashMap;
@@ -37,6 +37,8 @@ pub struct Renderer {
     pub option_as_alt: String,
     is_ime_enabled: bool,
     is_vi_mode_enabled: bool,
+    sugarloaf_content: ContentState,
+    sugarloaf_objects: Vec<Object>,
     pub is_kitty_keyboard_enabled: bool,
     pub last_typing: Option<Instant>,
     pub named_colors: Colors,
@@ -92,6 +94,8 @@ impl Renderer {
         }
 
         Renderer {
+            sugarloaf_content: ContentState::default(),
+            sugarloaf_objects: vec![],
             config_blinking_interval: config.cursor.blinking_interval.clamp(350, 1200),
             option_as_alt: config.option_as_alt.to_lowercase(),
             is_kitty_keyboard_enabled: config.keyboard.use_kitty_keyboard_protocol,
@@ -563,20 +567,6 @@ impl Renderer {
         }
     }
 
-    // #[inline]
-    // #[allow(dead_code)]
-    // fn create_graphic_sugar(&self, square: &Square) -> Sugar {
-    //     let media = &square.graphics().unwrap()[0].texture;
-    //     Sugar {
-    //         media: Some(SugarGraphic {
-    //             id: media.id,
-    //             width: media.width,
-    //             height: media.height,
-    //         }),
-    //         ..Sugar::default()
-    //     }
-    // }
-
     #[inline]
     fn create_cursor_style(&self, square: &Square) -> (FragmentStyle, char) {
         let font_attrs = match (
@@ -694,6 +684,12 @@ impl Renderer {
         self.is_vi_mode_enabled = is_vi_mode_enabled;
     }
 
+    #[inline]
+    pub fn prepare_from_last(&mut self, sugarloaf: &mut Sugarloaf) {
+        sugarloaf.set_content_state(self.sugarloaf_content.clone());
+        sugarloaf.set_objects(self.sugarloaf_objects.clone());
+    }
+
     #[allow(clippy::too_many_arguments)]
     pub fn prepare_term(
         &mut self,
@@ -731,7 +727,6 @@ impl Renderer {
 
         let content = sugarloaf.content();
 
-        // let start = std::time::Instant::now();
         for (i, row) in rows.iter().enumerate() {
             let has_cursor = is_cursor_visible && self.cursor.state.pos.row == i;
             self.create_line(
@@ -743,8 +738,8 @@ impl Renderer {
                 focused_match,
             );
         }
-        // let duration = start.elapsed();
-        // println!("Total loop rows: {:?}", duration);
+
+        self.sugarloaf_content = content.state.clone();
 
         let mut objects = Vec::with_capacity(30);
         self.navigation.build_objects(
@@ -765,6 +760,8 @@ impl Renderer {
 
             self.active_search = None;
         }
+
+        self.sugarloaf_objects = objects.clone();
 
         sugarloaf.set_objects(objects);
     }