+++ rio-backend/src/config/defaults.rs
@@ -16,15 +16,6 @@ pub fn default_cursor_interval() -> u64 {
 }
 
 #[inline]
-pub fn default_max_fps() -> u64 {
-    if cfg!(target_os = "macos") {
-        240
-    } else {
-        90
-    }
-}
-
-#[inline]
 pub fn default_padding_y() -> [f32; 2] {
     [0., 0.]
 }