+++ sugarloaf/src/lib.rs
@@ -19,5 +19,6 @@ pub use crate::sugarloaf::{
 };
 pub use components::rect::Rect;
 pub use layout::{
-    Content, FragmentStyle, FragmentStyleDecoration, UnderlineInfo, UnderlineShape,
+    Content, ContentState, FragmentStyle, FragmentStyleDecoration, UnderlineInfo,
+    UnderlineShape,
 };